Added creation of servers and channels. All kind of quality of life changes Bug fixes
