+++ Client/src/main/java/diskord/client/ChatFileType.java
@@ -1,6 +1,17 @@
 package diskord.client;
 
 public enum ChatFileType {
-    IMAGE,
-    FILE
+    IMAGE("IMAGE"),
+    FILE("FILE");
+
+    final String parameter;
+    ChatFileType(String parameter) {
+        this.parameter = parameter;
+    }
+
+    @Override
+    public String toString() {
+        return parameter;
+    }
+
 }
